% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfit.R
\name{change_digit}
\alias{change_digit}
\alias{sp_difference}
\title{Changes a single digit of a single parameter fit}
\usage{
change_digit(fit, which = "random")

sp_difference(original_fit, new_fit)
}
\arguments{
\item{fit, original_fit, new_fit}{fit objects returned by [sp_fit()]}

\item{which}{which decimal digit to change.}
}
\description{
Changes a single digit of a single parameter fit
}
\examples{
x <- rnorm(100)
fit <- sp_fit(x, binary_precision = 12)

# Randomly change the 30th decimal
new_fit <- change_digit(fit, 30)

# visualise the change (numbers in red)
sp_difference(fit, new_fit)

# visualise the impact on "predictions".
plot(x)
points(predict(fit), col = "red")
points(predict(new_fit), col = "blue")


}
