% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfit.R
\name{sp_fit}
\alias{sp_fit}
\title{Single Parameter Fit}
\usage{
sp_fit(data, binary_precision = 12)
}
\arguments{
\item{data}{numeric vector}

\item{binary_precision}{binary precision}
}
\description{
Encodes a numeric vector into a single parameter
}
\examples{
x <- rnorm(100)
fit <- sp_fit(x, binary_precision = 12)

# Print the single parameter
as.character(fit, digits = 20)

# Plot "observed" and "predicted"
plot(x)
points(predict(fit), col = "red")

# "Extrapolate" to unseen data
plot(predict(fit, length(x) + 10), col = "red")
points(x)

}
